# internal/static/swagger.json
{
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "The purpose of this service is to provide an application to help manage db switch task",
    "title": "DCSwitch API",
    "version": "0.0.1"
  },
  "host": "127.0.0.1",
  "paths": {
...
    "/{id}": {
      "get": {
        "description": "慢查询模拟",
        "operationId": "GetMockSlow",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Id",
            "description": "an id of user info",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GetMockSlowResponse"
          }
        }
      }
    }
  },
...
  "responses": {
...
    "GetMockSlowResponse": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Id"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "TOKEN",
      "in": "header"
    }
  },
  "security": [
    {
      "api_key": []
    }
  ]
}